requirements-dev.txt
-r requirements.txt
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
respx==0.22.0
